Fix circular dependency
