+++ packages/wallet-connectors/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+-   Fixed circular dependency between `index` and `WalletConnect` files.
+
 ### Added
 
 -   `WalletConnect`: `signMessage` now supports passing binary message including schema to the wallet.
